Base docker image off of Ubuntu 14.04
Ubuntu 16.04 has a newer version of libc.so, which results in compile
mismatch differences in checkbuild between the buildbot and RBE.
In fact, the build bot even produces a different result when compared
with corp workstations.

Digging further, I realized that this is because all the buildbots run
with ubuntu14.04 which has an older version of libc.so (2.19), which
ubuntu16.04 and onwards have higher versions.

An example compile where this happens is out/soong/.intermediates/external/vixl/libvixl/android_arm64_armv8-a_static/obj/external/vixl/src/aarch64/logic-aarch64.o

Bug: b/146347124